Voter Sortie

+++ src/Controller/SortieController.php
@@ -19,17 +19,17 @@ use Symfony\Component\Routing\Annotation\Route;
 #[Route('/sortie', name: 'sortie_')]
 class SortieController extends AbstractController
 {
-    #[Route('/{page}', name: 'show_all',  requirements: ["page"=>"\d+"])]
+    #[Route('/{page}', name: 'show_all', requirements: ["page" => "\d+"])]
     public function showAll(
-        SortieRepository $sortieRepository,
-        Request $request,
+        SortieRepository  $sortieRepository,
+        Request           $request,
         EtatSortieManager $etatSortieManager,
-        int $page = 1
-    ) : Response
+        int               $page = 1
+    ): Response
     {
 
         $nbSortieMax = count($sortieRepository->findAllToCheck());
-        $maxPage = ceil($nbSortieMax/SortieRepository::SORTIE_LIMIT);
+        $maxPage = ceil($nbSortieMax / SortieRepository::SORTIE_LIMIT);
 
         //Checker les etats
         $sorties = $sortieRepository->findAllToCheck();
@@ -41,44 +41,46 @@ class SortieController extends AbstractController
         $formFiltre->handleRequest($request);
 
 
-            if ($formFiltre->isSubmitted() && $formFiltre->isValid()) {
+        if ($formFiltre->isSubmitted() && $formFiltre->isValid()) {
 
-                $user = $this->getUser()->getId();
-                $sorties = $sortieRepository->findAllToDisplayFilter($model, $user);
+            $user = $this->getUser()->getId();
+            $sorties = $sortieRepository->findAllToDisplayFilter($model, $user);
 
-                return $this->render('sortie/showAll.html.twig', [
-                    'sorties' => $sorties,
-                    'filtreForm' => $formFiltre->createView(),
-                    "currentPage"=>1,
-                    "maxPage"=>1
-                ]);
-            }else {
-                if ($page >= 1 && $page <= $maxPage) {
+            return $this->render('sortie/showAll.html.twig', [
+                'sorties' => $sorties,
+                'filtreForm' => $formFiltre->createView(),
+                "currentPage" => 1,
+                "maxPage" => 1
+            ]);
+        } else {
+            if ($page >= 1 && $page <= $maxPage) {
 
                 $sorties = $sortieRepository->findAllToDisplay($page);
 
-                } else {
-                    throw $this->createNotFoundException("Oops ! Page non trouvée !");
-                }
-
+            } else {
+                throw $this->createNotFoundException("Oops ! Page non trouvée !");
             }
+
+        }
 
         return $this->render('sortie/showAll.html.twig', [
             'sorties' => $sorties,
             'filtreForm' => $formFiltre->createView(),
-            "currentPage"=>$page,
-            "maxPage"=>$maxPage
+            "currentPage" => $page,
+            "maxPage" => $maxPage
         ]);
     }
 
-    #[Route('/show/{id}', name: 'show_one', requirements:['id' => '\d+'])]
+    #[Route('/show/{id}', name: 'show_one', requirements: ['id' => '\d+'])]
     public function show(int $id, EtatSortieManager $etatSortieManager, SortieRepository $sortieRepository): Response
     {
 
         $id = $sortieRepository->findOneToDisplay($id);
         $id = $etatSortieManager->checkEtatSortie($id);
 
-        return $this->render('sortie/show.html.twig', ['sortie'=> $id]);
+        $this->denyAccessUnlessGranted('sortie_display', $id);
+
+        return $this->render('sortie/show.html.twig', ['sortie' => $id]);
     }
 
     #[Route('/add', name: 'add')]
@@ -95,7 +97,7 @@ class SortieController extends AbstractController
 
             if ($request->request->get('submit') == 1) {
                 $sortie->setEtat($etatRepository->find(1));
-            }elseif ($request->request->get('submit') == 2) {
+            } elseif ($request->request->get('submit') == 2) {
                 $sortie->setEtat($etatRepository->find(2));
             }
 
@@ -108,17 +110,18 @@ class SortieController extends AbstractController
 
         return $this->render(
             'sortie/add.html.twig',
-            ['sortieForm'=> $sortieForm->createView(), 'sortie'=>$sortie]
+            ['sortieForm' => $sortieForm->createView(), 'sortie' => $sortie]
         );
     }
 
-    #[Route('/update/{id}', name: 'update', requirements:['id' => '\d+'])]
+    #[Route('/update/{id}', name: 'update', requirements: ['id' => '\d+'])]
     public function update(SortieRepository $sortieRepository, Request $request, int $id): Response
     {
         $id = $sortieRepository->findOneToDisplay($id);
 
-        $sortieForm = $this->createForm(SortieType::class, $id);
+        $this->denyAccessUnlessGranted('sortie_update', $id);
 
+        $sortieForm = $this->createForm(SortieType::class, $id);
         $sortieForm->handleRequest($request);
 
         if ($sortieForm->isSubmitted() && $sortieForm->isValid()) {
@@ -130,14 +133,15 @@ class SortieController extends AbstractController
             return $this->redirectToRoute('sortie_show_one', ['id' => $id->getId()]);
         }
 
-        return $this->render('sortie/update.html.twig', ['sortieForm'=> $sortieForm->createView(),'sortie'=>$id]);
+        return $this->render('sortie/update.html.twig', ['sortieForm' => $sortieForm->createView(), 'sortie' => $id]);
     }
 
-    #[Route('/delete/{id}', name: 'delete', requirements:['id' => '\d+'])]
+    #[Route('/delete/{id}', name: 'delete', requirements: ['id' => '\d+'])]
     public function delete(SortieRepository $sortieRepository, Request $request, int $id): Response
     {
-
         $id = $sortieRepository->findOneToDisplay($id);
+
+        $this->denyAccessUnlessGranted('sortie_update', $id);
 
         if ($id) {
             $sortieRepository->remove($id, true);
@@ -150,44 +154,42 @@ class SortieController extends AbstractController
         return $this->redirectToRoute('main_home');
     }
 
-    #[Route('/cancel/{id}', name: 'cancel', requirements:['id' => '\d+'])]
+    #[Route('/cancel/{id}', name: 'cancel', requirements: ['id' => '\d+'])]
     public function cancel(
         SortieRepository $sortieRepository,
-        EtatRepository $etatRepository,
-        Request $request,
-        int $id
+        EtatRepository   $etatRepository,
+        Request          $request,
+        int              $id
     ): Response
     {
 
         $id = $sortieRepository->findOneToDisplay($id);
 
+        $this->denyAccessUnlessGranted('sortie_cancel', $id);
+
         $cancelForm = $this->createForm(CancelType::class, $id);
 
         $cancelForm->handleRequest($request);
 
-        if ($id->getOrganisateur() === $this->getUser() || $this->getUser()->getRoles()[0] == 'ROLE_ADMIN') {
+        if ($cancelForm->isSubmitted() && $cancelForm->isValid()) {
+            $id->setEtat($etatRepository->find(6));
 
-            if ($cancelForm->isSubmitted() && $cancelForm->isValid()) {
-                $id->setEtat($etatRepository->find(6));
+            $sortieRepository->save($id, true);
 
-                $sortieRepository->save($id, true);
+            $this->addFlash("warning", "Sortie annulée !");
 
-                $this->addFlash("warning", "Sortie annulée !");
-
-                return $this->redirectToRoute('main_home');
-
-            }
-            return $this->render('sortie/cancel.html.twig', ['cancelForm'=> $cancelForm->createView(),'sortie'=>$id]);
-
-        }else {
             return $this->redirectToRoute('main_home');
+
         }
+        return $this->render('sortie/cancel.html.twig', ['cancelForm' => $cancelForm->createView(), 'sortie' => $id]);
     }
 
-    #[Route('/publish/{id}', name: 'publish', requirements:['id' => '\d+'])]
+    #[Route('/publish/{id}', name: 'publish', requirements: ['id' => '\d+'])]
     public function publish(SortieRepository $sortieRepository, EtatRepository $etatRepository, int $id): Response
     {
         $id = $sortieRepository->findOneToDisplay($id);
+
+        $this->denyAccessUnlessGranted('sortie_update', $id);
 
         $id->setEtat($etatRepository->find(2));
 
@@ -195,49 +197,37 @@ class SortieController extends AbstractController
 
         $this->addFlash("success", "Sortie publiée !");
 
-        return $this->render('sortie/show.html.twig', ['sortie'=>$id]);
+        return $this->render('sortie/show.html.twig', ['sortie' => $id]);
     }
 
-    #[Route('/subscribe/{id}', name: 'subscribe', requirements:['id' => '\d+'])]
+    #[Route('/subscribe/{id}', name: 'subscribe', requirements: ['id' => '\d+'])]
     public function subscribe(SortieRepository $sortieRepository, int $id): Response
     {
         $id = $sortieRepository->findOneToDisplay($id);
 
-        $sortie = $sortieRepository->find($id);
-        if ($sortie->getEtat()->getLibelle()=='Ouverte') {
+        $this->denyAccessUnlessGranted('sortie_subscribe', $id);
 
-            $sortie->addParticipant($this->getUser());
-            $sortieRepository->save($sortie, true);
+        $id->addParticipant($this->getUser());
+        $sortieRepository->save($id, true);
 
-            $this->addFlash("primary", "Inscription pris en compte !");
+        $this->addFlash("primary", "Inscription pris en compte !");
 
-
-            $resultat = $this->render('sortie/show.html.twig', ['sortie' => $sortie]);
-
-        }else {
-
-            $resultat = $this->redirectToRoute('main_home');
-        }
-        return $resultat;
+        return $this->render('sortie/show.html.twig', ['sortie' => $id]);
     }
 
-    #[Route('/unsubscride/{id}', name: 'unsubscride', requirements:['id' => '\d+'])]
+    #[Route('/unsubscride/{id}', name: 'unsubscride', requirements: ['id' => '\d+'])]
     public function unsubscride(SortieRepository $sortieRepository, int $id): Response
     {
         $id = $sortieRepository->findOneToDisplay($id);
 
-        $sortie = $sortieRepository->find($id);
-    if ($sortie->getEtat()->getLibelle()=='Ouverte' || $sortie->getEtat()->getLibelle()=='Clôturée'){
-        $sortie->removeParticipant($this->getUser());
+        $this->denyAccessUnlessGranted('sortie_subscribe', $id);
 
-        $sortieRepository->save($sortie, true);
+        $id->removeParticipant($this->getUser());
+
+        $sortieRepository->save($id, true);
 
         $this->addFlash("danger", "Désinscription pris en compte !");
 
-        $resultat = $this->render('sortie/show.html.twig', ['sortie' => $sortie]);
-    }else {
-        $resultat = $this->redirectToRoute('main_home');
-    }
-        return $resultat;
+        return $this->render('sortie/show.html.twig', ['sortie' => $id]);
     }
 }
